#176 refactor: DTO 변환 메서드 Assembler -> DTO 이동 및 주석 추가
