+++ src/main/kotlin/com/psr/psr/inquiry/dto/InquiryListRes.kt
@@ -2,4 +2,11 @@ package com.psr.psr.inquiry.dto
 
 data class InquiryListRes (
     val inquiries: List<InquiryRes>?
-)
+) {
+    companion object {
+        fun toDto(inquiryList: List<InquiryRes>): InquiryListRes {
+            if (inquiryList.isEmpty()) return InquiryListRes(null)
+            return InquiryListRes(inquiryList)
+        }
+    }
+}
