+++ src/main/kotlin/com/psr/psr/notification/dto/Notification.kt
@@ -3,4 +3,13 @@ package com.psr.psr.notification.dto
 data class Notification(
     val title: String,
     val body: String
-)
+) {
+    companion object {
+        fun toDTO(title: String, body: String): Notification {
+            return Notification(
+                title = title,
+                body = body
+            )
+        }
+    }
+}
