+++ src/main/kotlin/com/psr/psr/order/dto/OrderRes.kt
@@ -1,5 +1,8 @@
 package com.psr.psr.order.dto
 
+import com.psr.psr.order.entity.Order
+import java.time.format.DateTimeFormatter
+
 data class OrderRes (
     val isSeller: Boolean,
     val status: String,
@@ -11,4 +14,21 @@ data class OrderRes (
     var websiteUrl: String? = null,
     val inquiry: String,
     val description: String
-)
+) {
+    companion object {
+        fun toOrderResDTO(order: Order, isSeller: Boolean): OrderRes {
+            return OrderRes(
+                isSeller = isSeller,
+                status = order.orderStatus.value,
+                orderUserId = order.user.id!!,
+                orderDate = order.createdAt.format(DateTimeFormatter.ISO_DATE),
+                productId = order.product.id!!,
+                productName = order.product.name,
+                ordererName = order.ordererName,
+                websiteUrl = order.websiteUrl,
+                inquiry = order.inquiry,
+                description = order.description
+            )
+        }
+    }
+}
