+++ src/main/kotlin/com/psr/psr/review/dto/ReviewRes.kt
@@ -1,5 +1,7 @@
 package com.psr.psr.review.dto
 
+import com.psr.psr.review.entity.Review
+
 data class ReviewRes(
     val reviewId: Long,
     val rating: Int,
@@ -8,4 +10,21 @@ data class ReviewRes(
     val nickname: String,
     val productName: String,
     val productImgUrl: String?
-)
+) {
+    companion object {
+        fun toDto(review: Review): ReviewRes {
+            val reviewImgs =
+                if (review.imgs.isNotEmpty()) review.imgs.map { img -> img.imgUrl }
+                else null
+            return ReviewRes(
+                reviewId = review.id!!,
+                rating = review.rating,
+                content = review.content,
+                imgList = reviewImgs,
+                nickname = review.product.user.nickname,
+                productName = review.product.name,
+                productImgUrl = review.product.imgs.firstOrNull()?.imgUrl
+            )
+        }
+    }
+}
