+++ src/main/kotlin/com/psr/psr/review/entity/Review.kt
@@ -38,4 +38,15 @@ data class Review(
         this.content = reviewReq.content
         return this
     }
+
+    companion object {
+        fun toEntity(order: Order, reviewReq: ReviewReq): Review {
+            return Review(
+                order = order,
+                product = order.product,
+                rating = reviewReq.rating,
+                content = reviewReq.content
+            )
+        }
+    }
 }
